fixes #5605 - ignore cases where output record is null
This commit is to address a case that we observed with
a couple of the hammer-cli-katello commands.

E.g.
command 'activation-key list' has the following output
defined:

        from :environment do
          field :name, _("Lifecycle Environment")
        end
        from :content_view do
          field :name, _("Content View")
        end

Since environment and content_view are optional, there
are cases where this would result in a 'null' back
from the server for those entities.  When that occurs
the following error would be generated to the user:

Error: undefined method `has_key?' for nil:NilClass

This small change addresses that error.