Fixes #17252 - Allow empty parameters in hammer
This change allows user to use nil as a valid value for an argument.

  $ hammer command --option=NIL
  $ HAMMER_NIL=NULL hammer command --option=NULL

Just before option normalization the 'NIL' string is replaced for
special internal object which help us to distinguish between unused
options and options set to nil. The internal object is replaced with
nill fter all options are collected and unused options filtered out..